[LibOS] Do not try to send signal 0 in kill()/tkill()
Previously, Graphene didn't handle special case of kill()/tkill()
with signal equal to zero. This signal is used to check for existence
of processes and threads. However, Graphene tried to send/append it
which led to segfaults. This commit correctly handles signal equal
to zero and improves LibOS tests to verify this fix.

+++ LibOS/shim/test/regression/signal_multithread.c
@@ -51,6 +51,12 @@ int main() {
 
     wait_for(1);
 
+    /* the below dummy kill (no signal is sent) is for sanity */
+    if (kill(getpid(), /*sig=*/0)) {
+        fprintf(stderr, "kill(sig=0) failed: %m\n");
+        return 1;
+    }
+
     if (kill(getpid(), SIGTERM)) {
         fprintf(stderr, "kill failed: %m\n");
         return 1;
